add timer.css to decompose module
