+++ src/modules/timer-module/timer.css
@@ -1,0 +1,32 @@
+.timer {
+    background-color: #343434;
+    color: #eee;
+    display: none;
+    flex-direction: column;
+    width: 250px;
+    justify-content: space-between;
+    border-radius: 20px;
+    padding: 10px;
+    position: fixed;
+    z-index: 1000;
+  }
+  
+  .timer.active {
+    display: flex;
+  }
+  
+  .timer .timer__input {
+    width: 100%;
+    height: 40px;
+    border-radius: 10px;
+    padding: 5px;
+    margin: 10px auto;
+  }
+  
+  .timer .timer__button {
+    width: 100%;
+    height: 40px;
+    border-radius: 10px;
+    padding: 5px;
+    margin: 10px auto;
+  }
