+++ src/styles.css
@@ -40,35 +40,3 @@ body {
 .menu .menu-item:hover {
   background: #343434;
 }
-
-.timer {
-  background-color: #343434;
-  color: #eee;
-  display: none;
-  flex-direction: column;
-  width: 250px;
-  justify-content: space-between;
-  border-radius: 20px;
-  padding: 10px;
-  position: fixed;
-}
-
-.timer.active {
-  display: flex;
-}
-
-.timer .timer__input {
-  width: 100%;
-  height: 40px;
-  border-radius: 10px;
-  padding: 5px;
-  margin: 10px auto;
-}
-
-.timer .timer__button {
-  width: 100%;
-  height: 40px;
-  border-radius: 10px;
-  padding: 5px;
-  margin: 10px auto;
-}
